+++ gulpfile.js
@@ -1,5 +1,5 @@
 (function(){
-    'use strict';    
+    'use strict';
     var gulp = require('gulp'),
         gutil = require('gulp-util'),
         connect = require('gulp-connect'),
@@ -51,6 +51,9 @@
                 'src/js/keyboard.js',
                 'src/js/mousewheel.js',
                 'src/js/parallax.js',
+                'src/js/plugins.js',
+                'src/js/emitter.js',
+                'src/js/a11y.js',
                 'src/js/init.js',
                 'src/js/swiper-outro.js',
                 'src/js/swiper-proto.js',
@@ -71,6 +74,9 @@
                 'src/js/keyboard.js',
                 'src/js/mousewheel.js',
                 'src/js/parallax.js',
+                'src/js/plugins.js',
+                'src/js/emitter.js',
+                'src/js/a11y.js',
                 'src/js/init.js',
                 'src/js/swiper-outro.js',
                 'src/js/swiper-proto.js',
@@ -79,13 +85,16 @@
                 'src/js/amd.js'
             ],
             Framework7Files : [
-                'src/js/swiper-intro.js',
+                'src/js/swiper-intro-f7.js',
                 'src/js/core.js',
                 'src/js/effects.js',
                 'src/js/lazy-load.js',
                 'src/js/scrollbar.js',
                 'src/js/controller.js',
                 'src/js/parallax.js',
+                'src/js/plugins.js',
+                'src/js/emitter.js',
+                'src/js/a11y.js',
                 'src/js/init.js',
                 'src/js/swiper-outro.js',
                 'src/js/swiper-proto.js',
@@ -113,14 +122,14 @@
                 day: new Date().getDate()
             }
         };
-        
+
     function addJSIndent (file, t, minusIndent) {
         var addIndent = '        ';
         var filename = file.path.split('src/js/')[1];
         if (filename === 'wrap-start.js' || filename === 'wrap-end.js' || filename === 'amd.js') {
             addIndent = '';
         }
-        if (filename === 'swiper-intro.js' || filename === 'swiper-outro.js' || filename === 'dom.js' || filename === 'dom-plugins.js' || filename === 'swiper-proto.js') addIndent = '    ';
+        if (filename === 'swiper-intro.js' || filename === 'swiper-intro-f7.js' || filename === 'swiper-outro.js' || filename === 'dom.js' || filename === 'dom-plugins.js' || filename === 'swiper-proto.js') addIndent = '    ';
         if (minusIndent) {
             addIndent = addIndent.substring(4);
         }
@@ -138,7 +147,7 @@
             .pipe(tap(function (file, t){
                 addJSIndent (file, t);
             }))
-            .pipe(concat(swiper.filename + '.js'))            
+            .pipe(concat(swiper.filename + '.js'))
             .pipe(header(swiper.banner, { pkg : swiper.pkg, date: swiper.date } ))
             .pipe(gulp.dest(paths.build.scripts))
 
@@ -215,7 +224,10 @@
 
         gulp.src(paths.build.styles + '*.css')
             .pipe(gulp.dest(paths.dist.styles))
-            .pipe(minifyCSS())
+            .pipe(minifyCSS({
+                advanced: false,
+                aggressiveMerging: false,
+            }))
             .pipe(header(swiper.banner, { pkg : swiper.pkg, date: swiper.date }))
             .pipe(rename(function(path) {
                 path.basename = swiper.filename + '.min';
@@ -235,7 +247,7 @@
             port:'3000'
         });
     });
-    
+
     gulp.task('open', function () {
         return gulp.src(paths.playground.root + 'index.html').pipe(open('', { url: 'http://localhost:3000/' + paths.playground.root + 'index.html'}));
     });