+++ CHANGELOG.md
@@ -1,5 +1,45 @@
 # Change Log
 
+## Swiper 3.0.7 - Released on April 25, 2015
+  * New `width` and `height` parameters to force Swiper size, useful when it is hidden on intialization
+  * Better support for "Scroll Container". So now Swiper can be used as a scroll container with one single "scrollable"/"swipeable" slide
+  * Added lazy loading for background images with `data-background` attribute on required elements
+  * New "Sticky Free Mode" (with `freeModeSticky` parameter) which will snap to slides positions in free mode
+  * Fixed issues with lazy loading  
+  * Fixed slide removing when loop mode is enabled
+  * Fixed issues with Autoplay and Fade effect
+  * Minor fixes
+
+## Swiper 3.0.6 - Released on March 27, 2015
+  * Fixed sometimes wrong slides position when using "Fade" effect
+  * `.destroy(deleteInstance, cleanupStyles)` method now has second `cleanupStyles` argument, when passed - all custom styles will be removed from slides, wrapper and container. Useful if you need to destroy Swiper and to init again with new options or in different direction
+  * Minor fixes
+
+## Swiper 3.0.5 - Released on March 21, 2015
+  * New Keyboard accessibility module to provide foucsable navigation buttons and basic ARIA for screen readers with new parameters:
+    * `a11y: false` - enable accessibility
+    * `prevSlideMessage: 'Previous slide'` - message for screen readers for previous button
+    * `nextSlideMessage: 'Next slide'` - message for screen readers for next button
+    * `firstSlideMessage: 'This is the first slide'` - message for screen readers for previous button when swiper is on first slide
+    * `lastSlideMessage: 'This is the last slide'` - message for screen readers for next button when swiper is on last slide
+  * New Emitter module. It allows to work with callbacks like with events, even adding them after initialization with new methods:
+    * `.on(event, handler)` - add event/callback
+    * `.off(event, handler)` - remove this event/callback
+    * `.once(event, handler)` - add event/callback that will be executed only once
+  * Plugins API is back. It allows to write custom Swiper plugins
+  * Better support for browser that don't support flexbox layout
+  * New parameter `setWrapperSize` (be default it is `false`) to provide better compatibility with browser without flexbox support. Enabled this option and plugin will set width/height on swiper wrapper equal to total size of all slides
+  * New `virtualTranslate` parameter. When it is enabled swiper will be operated as usual except it will not move. Useful when you may need to create custom slide transition
+  * Added support for multiple Pagination containers
+  * Fixed `onLazyImage...` callbacks
+  * Fixed issue with not accessible links inside of Slides on Android < 4.4
+  * Fixed pagination bullets behavior in loop mode with specified `slidesPerGroup`
+  * Fixed issues with clicks on IE 10+ touch devices
+  * Fixed issues with Coverflow support on IE 10+
+  * Hashnav now will update document hash after transition to prevent browsers UI lags, not in the beginning like before
+  * Super basic support for IE 9 with swiper.jquery version. No animation and transitions, but basic stuff like switching slides/pagination/scrollbars works
+  
+
 ## Swiper 3.0.4 - Released on March 6, 2015
   * New Images Lazy Load component
     * With new parameters `lazyLoading`, `lazyLoadingInPrevNext`, `lazyLoadingOnTransitionStart` (all disabled by default)